requirements.txt
anthropic
yfinance
ccxt
alpaca-py
apscheduler
pandas
ta
python-dotenv
pytest
pytest-mock
